revamp: using clean Architecture,
not tested yet, do not merge into master
