+++ lib/core/data/data_source/live_stream/live_stream_data_source.dart
@@ -1,27 +1,24 @@
-import 'package:flutter/cupertino.dart';
 import 'package:srk_monitor/core/data/http_services/http_services.dart';
 
+import '../../api/api.dart';
 import '../../dto/live_stream/live_url_dto.dart';
-import '../../../../views/live_stream/utils/api.dart';
 
 class LiveStreamDataSource {
   static LiveStreamDataSource? _instance;
   factory LiveStreamDataSource() => _instance ??= LiveStreamDataSource._();
   LiveStreamDataSource._();
 
-  // TODO next step: create providers
-  Future<List<String>?> getStreamUrlList(String roomId) async {
+  Future<LiveStreamUrlDto?> getStreamUrlDto(
+    String roomId, {
+    required bool useM3u8,
+  }) async {
     try {
-      final String url = LiveStreamApi.getLiveStreamUrl(roomId);
+      final String url = Api.getLiveStreamUrl(roomId, useM3u8);
       final jsonResponse = await HttpService().get(url);
       LiveStreamUrlDto response = LiveStreamUrlDto.fromJson(jsonResponse);
-      List<String> urlList = [];
-      response.data?.durl?.forEach((element) {
-        urlList.add(element.url ?? '');
-      });
-      return urlList;
+      return response;
     } catch (e) {
-      debugPrint('LiveStreamServices.getStreamUrlList error:  $e');
+      return null;
     }
   }
 }
