+++ lib/core/data/dto/home_page/room_info_dto.dart
@@ -111,42 +111,42 @@ class RoomInfoDto {
   }
 
   Map<String, dynamic> toJson() {
-    final _data = <String, dynamic>{};
-    _data['uid'] = uid;
-    _data['room_id'] = roomId;
-    _data['short_id'] = shortId;
-    _data['attention'] = attention;
-    _data['online'] = online;
-    _data['is_portrait'] = isPortrait;
-    _data['description'] = description;
-    _data['live_status'] = liveStatus;
-    _data['area_id'] = areaId;
-    _data['parent_area_id'] = parentAreaId;
-    _data['parent_area_name'] = parentAreaName;
-    _data['old_area_id'] = oldAreaId;
-    _data['background'] = background;
-    _data['title'] = title;
-    _data['user_cover'] = userCover;
-    _data['keyframe'] = keyframe;
-    _data['is_strict_room'] = isStrictRoom;
-    _data['live_time'] = liveTime;
-    _data['tags'] = tags;
-    _data['is_anchor'] = isAnchor;
-    _data['room_silent_type'] = roomSilentType;
-    _data['room_silent_level'] = roomSilentLevel;
-    _data['room_silent_second'] = roomSilentSecond;
-    _data['area_name'] = areaName;
-    _data['pendants'] = pendants;
-    _data['area_pendants'] = areaPendants;
-    _data['hot_words'] = hotWords;
-    _data['hot_words_status'] = hotWordsStatus;
-    _data['verify'] = verify;
-    _data['up_session'] = upSession;
-    _data['pk_status'] = pkStatus;
-    _data['pk_id'] = pkId;
-    _data['battle_id'] = battleId;
-    _data['allow_change_area_time'] = allowChangeAreaTime;
-    _data['allow_upload_cover_time'] = allowUploadCoverTime;
-    return _data;
+    final data = <String, dynamic>{};
+    data['uid'] = uid;
+    data['room_id'] = roomId;
+    data['short_id'] = shortId;
+    data['attention'] = attention;
+    data['online'] = online;
+    data['is_portrait'] = isPortrait;
+    data['description'] = description;
+    data['live_status'] = liveStatus;
+    data['area_id'] = areaId;
+    data['parent_area_id'] = parentAreaId;
+    data['parent_area_name'] = parentAreaName;
+    data['old_area_id'] = oldAreaId;
+    data['background'] = background;
+    data['title'] = title;
+    data['user_cover'] = userCover;
+    data['keyframe'] = keyframe;
+    data['is_strict_room'] = isStrictRoom;
+    data['live_time'] = liveTime;
+    data['tags'] = tags;
+    data['is_anchor'] = isAnchor;
+    data['room_silent_type'] = roomSilentType;
+    data['room_silent_level'] = roomSilentLevel;
+    data['room_silent_second'] = roomSilentSecond;
+    data['area_name'] = areaName;
+    data['pendants'] = pendants;
+    data['area_pendants'] = areaPendants;
+    data['hot_words'] = hotWords;
+    data['hot_words_status'] = hotWordsStatus;
+    data['verify'] = verify;
+    data['up_session'] = upSession;
+    data['pk_status'] = pkStatus;
+    data['pk_id'] = pkId;
+    data['battle_id'] = battleId;
+    data['allow_change_area_time'] = allowChangeAreaTime;
+    data['allow_upload_cover_time'] = allowUploadCoverTime;
+    return data;
   }
 }
