+++ lib/core/domain/entity/live_stream/live_stream_entity.dart
@@ -1,0 +1,6 @@
+class LiveStreamEntity {
+  LiveStreamEntity({
+    required this.urls,
+  });
+  List<String> urls;
+}
