+++ lib/core/domain/repository/home_page/home_page_repository.dart
@@ -5,24 +5,13 @@ import 'package:srk_monitor/core/data/dto/base_resopnse.dart';
 import 'package:srk_monitor/core/data/dto/home_page/live_user_dto.dart';
 import 'package:srk_monitor/core/data/dto/home_page/room_info_dto.dart';
 
-import '../../entity/home_page/home_grid_entity.dart';
-
 class HomePageRepository {
   final LiveRoomInfoDataSource _liveRoomInfoDataSource =
       LiveRoomInfoDataSource();
   final LiveUserInfoDataSource _liveUserInfoDataSource =
       LiveUserInfoDataSource();
 
-  Future<HomeGridEntity> getHomePageLiveDetails(String roomId) async {
-    final RoomInfoDto? roomInfoDto = await _getRoomInfoDto(roomId);
-    late final LiveUserDto? liveUserDto;
-    if (roomInfoDto != null) {
-      liveUserDto = await _getLiveUserDto(roomInfoDto.uid.toString());
-    }
-    // TODO Mapper;
-  }
-
-  Future<LiveUserDto?> _getLiveUserDto(String uid) async {
+  Future<LiveUserDto?> getLiveUserDto(String uid) async {
     try {
       final BaseResponse response =
           await _liveUserInfoDataSource.getLiveUser(uid);
@@ -30,12 +19,14 @@ class HomePageRepository {
         LiveUserDto liveUserDto = response.data;
         return liveUserDto;
       }
+      return null;
     } catch (e) {
       debugPrint('HomePageRepository._getLiveUserDto error: $e');
+      return null;
     }
   }
 
-  Future<RoomInfoDto?> _getRoomInfoDto(String roomId) async {
+  Future<RoomInfoDto?> getRoomInfoDto(String roomId) async {
     try {
       final BaseResponse response =
           await _liveRoomInfoDataSource.getRoomInfo(roomId);
@@ -43,8 +34,10 @@ class HomePageRepository {
         RoomInfoDto roomInfoDto = response.data;
         return roomInfoDto;
       }
+      return null;
     } catch (e) {
       debugPrint('HomePageRepository._getRoomInfoDto error: $e');
+      return null;
     }
   }
 }
