+++ lib/feature/views/live_stream/widget/live_page_items.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+import '../preset_page/small_9.dart';
+
+class LivePageItem extends StatelessWidget {
+  const LivePageItem({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Small9();
+  }
+}
