+++ lib/feature/views/setting_page/widgets/animating_view.dart
