+++ lib/main.dart
@@ -5,9 +5,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:srk_monitor/constants/colors.dart';
 import 'package:window_manager/window_manager.dart';
 
-import 'views/home_page/home_layout.dart';
-import 'views/setting_page/settings_items/general.dart';
-import 'views/setting_page/settings_layout.dart';
+import 'feature/views/home_page/home_layout.dart';
+import 'feature/views/setting_page/settings_items/general.dart';
+import 'feature/views/setting_page/settings_layout.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
