--- lib/views/live_stream/utils/api.dart
@@ -1,6 +1,0 @@
-class LiveStreamApi {
-  static getLiveStreamUrl(String roomid) {
-    String url =
-        "https://api.live.bilibili.com/room/v1/Room/playUrl?build=5470400&device=android&cid=$roomid";
-  }
-}
