--- lib/views/live_stream/widget/live_page_items.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../preset_page/small_9.dart';
-
-class LivePageItem extends StatelessWidget {
-  const LivePageItem({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return small_9();
-  }
-}
