--- lib/views/setting_page/widgets/settings_view.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:srk_monitor/views/setting_page/widgets/settings_items.dart';
-
-class SettingsView extends StatelessWidget {
-  final double padding = 448;
-  const SettingsView({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    double width = MediaQuery.of(context).size.width;
-    int crossAxisCount = (width - padding) ~/ 160;
-
-    return const SettignsItems();
-  }
-}
